chunk sender w +1 client

+++ internal/gateway/id_generator/id_generator.go
@@ -4,7 +4,7 @@ import "strconv"
 
 type IdGenerator struct {
 	prefix uint8
-	nextId uint64
+	nextId uint16
 }
 
 func New(prefix uint8) *IdGenerator {
